fix(sera-runtime): pre-dispatch CapabilityPolicy enforcement (sera-eo71)
Move tool-call CapabilityPolicy enforcement out of the gateway's post-hoc
event-rewriting path and into sera-runtime's tool dispatcher so a denied
tool truly never starts (no child process, no handler invocation).

- Move `CapabilityRegistry`, `CapabilityPolicy`, `PolicyDenial`, and
  `CapabilityRegistryError` into `sera-config::capability_registry`. The
  gateway's `capability_enforcement` module shrinks to a `pub use`
  re-export (so existing imports compile) plus an updated module doc.
- Load the registry once in `sera-runtime/src/main.rs` at startup. The
  gateway forwards the agent's `policyRef` (and any
  `SERA_CAPABILITY_POLICIES_DIR` override) to the runtime via env vars so
  the single-agent runtime can bind itself.
- `RegistryDispatcher::with_capability_registry(registry, agent_id)`
  attaches the registry to the dispatcher; the first thing `dispatch`
  does on every call is `registry.check(agent_id, tool)`. A deny returns
  `ToolError::PolicyDenied` *before* hooks, escalation, and
  `TraitToolRegistry::execute` — i.e. the tool handler is never invoked.
- Drop the gateway-side rewriting of `tool_call_begin/end` events. The
  gateway still observes incoming events and emits OCSF Policy Activity
  audit entries on observed denials (defence-in-depth).

Tests:
- Three new dispatcher unit tests in `sera-runtime/src/tools/dispatcher.rs`:
  `dispatcher_capability_deny_blocks_handler_pre_dispatch` (asserts the
  handler is never reached on deny — uses a non-existent tool name so a
  reached handler would surface as `NotFound` instead of `PolicyDenied`),
  `dispatcher_capability_allow_runs_tool_normally`, and
  `dispatcher_capability_unbound_agent_is_permissive`.
- Existing tests for the registry semantics moved to `sera-config` along
  with the type and continue to pass unchanged.

Deviations:
- Hot-reload is not implemented in v1 on the runtime side: the gateway
  still has `/admin/policies/reload` for its own registry (used by the
  audit pass), but the runtime child does not yet observe that swap.
  Operators restart the runtime to pick up policy changes.

Verification:
- cargo build --workspace --bins: pass
- cargo clippy --workspace -- -D warnings: pass
- cargo test --workspace: 3821 passed, 17 ignored, 0 failures

Closes sera-eo71.
